Add turn-based combat: hit/damage resolution, monster generation, attack/flee tools
Redesigns the ability model (SupportAbility split into instant heal/HoT/buff,
conditions split into DoT/status) to remove nullable-field ambiguity, adds
HitCalculator (proficiency-vs-defense hit chance, D2-style physical block) and
DamageCalculator (weapon-roll physical vs self-contained magic damage, both
mitigated by matching resistance), and rescales ability damage to a
percent-of-weapon-damage scale accordingly.

Adds MonsterGenerator to populate dungeon rooms with hostile creatures, and
wires everything up to the LLM via new AttackTool/FleeTool game tools: lazy
encounter start (pulls in every nearby hostile, not just the named target),
weapon-proficiency growth from swings, and combat rewards (XP/gold) persisted
back to the player on victory. CombatState.WeaponSwingCounts stays off the
wire; only a narrow CombatResult reaches the model.

Also fixes several latent scale/tracking bugs found while building this out:
chance fields (BlockChance, proc/special-hit chances) standardized to float
[0,1], CreatureWeaponProficiency given a real composite primary key, and
SetWeaponProficiencyCommand converted from a no-op-on-missing-row update to a
proper upsert.

+++ TRPG.Data/Models/ShieldItem.cs
@@ -2,7 +2,7 @@ namespace TRPG.Data.Models;
 
 public class ShieldItem : Item
 {
-    public int BlockChance { get; init; }
+    public float BlockChance { get; init; }
     public override EquipmentSlot? DefaultSlot => EquipmentSlot.LeftHand;
     public int Defense { get; init; }
     public int DurabilityCurrent { get; set; }
